a1: quality polish -- defined aliasing, clang parity leg, -Wconversion decoder, analyzer-clean
- ARENA overlays (SA apply state, u32 literal-seed histograms) now go through
  __attribute__((may_alias)) types: defined behavior under ANY integrator's
  flags, not just -O2-doesn't-happen-to-break. Documented at the SAst site.
- clang parity: building with clang exposed that the 640 B gcc-sized coroutine
  stack SEGFAULTS under clang codegen (measured high-water 3752 B vs gcc's
  520 B -- compiler-specific frame sizes). DEC_STACK_BYTES default is now
  per-compiler (gcc 640 / clang 4096) with a mandatory-re-measure note
  (RC_V3_STACKMEAS) for any compiler/flag change. PULL mode is unaffected (no
  coroutine stack). CI gains a clang leg: warning-free build + check +
  malformed + golden -- clang's encoder emits byte-identical blobs, proving
  compiler-independent wire.
- Decoder TU now builds -Wconversion-clean under gcc AND clang (9 implicit
  narrowing/sign sites given explicit casts, no behavior change); the flag is
  pinned in the hy_dec/hy_dec_pull rules so it stays clean.
- gcc -fanalyzer runs clean over both TUs (no gate wiring needed -- zero
  findings to guard).

Wire unchanged: golden manifest OK, corpus 4199788, one-face 589/305.

+++ src/patch_apply_demo.c
@@ -78,7 +78,7 @@ int main(int argc,char**argv){
     FILE*bf=fopen(argv[2],"rb"); if(!bf){perror("blob");return 2;}
     fseek(bf,0,SEEK_END); long bsz=ftell(bf); fseek(bf,0,SEEK_SET);
     if(bsz<12){ fprintf(stderr,"blob too short\n"); fclose(bf); return 1; }
-    uint8_t*blob=malloc(bsz); if(fread(blob,1,bsz,bf)!=(size_t)bsz){ fclose(bf); return 2; } fclose(bf);
+    uint8_t*blob=malloc((size_t)bsz); if(fread(blob,1,(size_t)bsz,bf)!=(size_t)bsz){ fclose(bf); return 2; } fclose(bf);
     /* MINIMAL envelope pre-parse, for the HOST EMULATOR ONLY (flash sizing + memfile sanity).
      * The decoder performs the authoritative envelope parse and both CRC gates itself; a device
      * integration needs none of this — its flash is fixed hardware.
